rna: Split CNV-expression correlation calculation into separate script
Add the TCGA melanoma (SKCM) correlation coefficients table to data/.
Update CLI options in the two import* scripts to take the correlation
table instead of the raw CNV and expression tables as input.
